webserv: Logger ref 5
