+++ headers/ClientConnection.h
