+++ headers/raii_ptr.h
