+++ srcs/logger/logUtils.cpp
@@ -1,16 +1,16 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   loggerUtils.cpp                                    :+:      :+:    :+:   */
+/*   logUtils.cpp                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: cflorind <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/12 12:23:58 by cflorind          #+#    #+#             */
-/*   Updated: 2022/04/14 12:02:33 by cflorind         ###   ########.fr       */
+/*   Updated: 2022/04/14 16:06:40 by cflorind         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "Logger/loggerUtils.h"
+#include "logger/logUtils.h"
 
 std::string	time_stamp(void)
 {
